% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edma_scale.R
\name{edma_scale}
\alias{edma_scale}
\title{Scale an EDMA data object}
\usage{
edma_scale(x, scale_by, L1 = NULL, L2 = NULL, scale_constant = NULL)
}
\arguments{
\item{x}{an EDMA data object of class 'edma_data'.}

\item{scale_by}{string specifying the type of scaling. Valid options are
"constant", "endpoints", "geometric mean", "maximum", "median", "sneath".
See below for details.}

\item{L1}{string specifying first landmark to use if scale_by = "endpoints"}

\item{L2}{string specifying second landmark to use if scale_by = "endpoints"}

\item{scale_constant}{numeric specifying the scaling constant to use for
scale_by = "constant"}
}
\value{
object of class 'edma_data', with landmarks scaled according to
scale_by parameter. See details for details of scaling procedures. The
object x are appended the scaling method string and the values used for
scaling. The latter can be useful for comparisons, e.g., of geometric
means.
}
\description{
This function implements the landmark scaling procedures described in
Lele and Cole (1996) which are used to rescale the landmarks for
specimens in which the variance-covariance matrices to two populations
are unequal.
}
\details{
\code{constant} Interlandmark distances are scaled by a numeric constant
that is applied to all specimens.

\code{endpoints} Interlandmark distances are scaled by the distance between
a pair of landmarks (\code{L1} and \code{L2}) for each specimen.

\code{geometric mean} Interlandmark distances are scaled by the geometric
mean of all pairwise distances for each specimen.

\code{maximum} Interlandmark distances are scaled by the maximum of all
pairwise distances for each specimen.

\code{median} Interlandmark distances are scaled by the median of all
pairwise distances for each specimen.

\code{sneath}  Interlandmark distances are scaled using the method described
by Sneath (1967), which uses the square-root of the mean squared
distances of each landmark to the centroid.
}
\examples{
# Following example in Lele and Cole (1996)
X <- matrix(c(0, 0, 2, 3, 4, 1), byrow = TRUE, ncol = 2)
Y <- matrix(c(0, 0, 3, 3, 3, 0), byrow = TRUE, ncol = 2)

XY <- as.edma_data(array(dim = c(3, 2, 2),
                         data = cbind(X, Y)))

XY_const <- edma_scale(XY, scale_by = "constant", scale_constant = 2)
XY_const$data
XY_const$scaling_method
XY_const$scaling_values

XY_endpt <- edma_scale(XY, scale_by = "endpoints", L1 = "L1", L2 = "L3")
XY_endpt$data
XY_endpt$scaling_method
XY_endpt$scaling_values

XY_geomean <- edma_scale(XY, scale_by = "geometric mean")
XY_geomean$data
XY_geomean$scaling_method
XY_geomean$scaling_values

XY_max <- edma_scale(XY, scale_by = "maximum")
XY_max$data
XY_max$scaling_method
XY_max$scaling_values

XY_median <- edma_scale(XY, scale_by = "median")
XY_median$data
XY_median$scaling_method
XY_median$scaling_values

XY_sneath <- edma_scale(XY, scale_by = "sneath")
XY_sneath$data
XY_sneath$scaling_method
XY_sneath$scaling_values

}
\references{
Lele, S., and T. M. Cole III. 1996. A new test for shape
differences when variance-covariance matrices are unequal.
J. Hum. Evol. 31:193-212.

Sneath, P. H. A. 1967. Trend-surface analysis of transformation grids.
J. Zool. 151:65-122. Wiley.
}
