\name{EDMAinR-package}
\alias{EDMAinR-package}
\alias{EDMAinR}
\docType{package}
\title{
\packageTitle{EDMAinR}
}
\description{
\packageDescription{EDMAinR}
}
\details{
EDMA data: \code{\link{read_xyz}}

Nonparametric fit: \code{\link{edma_fit}}

Form difference: \code{\link{edma_fdm}}

Growth and growth difference:
\code{\link{edma_gm}}, \code{\link{edma_gdm}}

Shape difference: \code{\link{edma_sdm}}
}
\author{
\packageAuthor{EDMAinR}

Maintainer: \packageMaintainer{EDMAinR}
}
\references{
Lele, S. R., 1991.
Some comments on coordinate-free and scale-invariant
methods in morphometrics.
American Journal of Physical Anthropology 85:407--417.
<doi:10.1002/ajpa.1330850405>

Lele, S. R., and Richtsmeier, J. T., 1991.
Euclidean distance matrix analysis: A coordinate-free approach for
comparing biological shapes using landmark data.
American Journal of Physical Anthropology 86(3):415--27.
<doi:10.1002/ajpa.1330860307>

Lele, S. R., and Richtsmeier, J. T., 1992.
On comparing biological shapes: detection of influential landmarks.
American Journal of Physical Anthropology 87:49--65.
<doi:10.1002/ajpa.1330870106>

Lele, S. R., and Richtsmeier, J. T., 1995.
Euclidean distance matrix analysis: confidence intervals for
form and growth differences.
American Journal of Physical Anthropology 98:73--86.
<doi:10.1002/ajpa.1330980107>

Hu, L., 2007. Euclidean Distance Matrix Analysis of Landmarks Data:
Estimation of Variance. Thesis, Master of Science in Statistics,
Department of Mathematical and Statistical Sciences,
University of Alberta, Edmonton, Alberta, Canada. Pp. 49.

Lele, S. R., and Cole, T. M. III., 1996.
A new test for shape differences when
variance-covariance matrices are unequal.
Journal of Human Evolution 31:193--212.
<doi:10.1006/jhev.1996.0057>
}
\keyword{ package }
